Changes necessary to satisfy dash and mawk
Remove bashisms in parameter expansions and replace with POSIX
alternatives. Old mawk 1.3.3 (12 years old at this point) is (still)
the default on Debian and Ubuntu, and is not POSIX because it
does not support character classes. Replaced with ASCII equivalents
for a compromise. Also, make sure there is always something to shift
when doing a shift, as some shells (dash) will complain when shifting
an empty argument list.