Major updates and improvements (#17)
The major change here is only pushing the specified directory when pushconfig is used with VCS support. Previously it was pushing the whole repository. Anyone on the old version might need to modify their TFE workspace's TF_ATLAS_DIR setting appropriately. If their config was at the root of the repo anyway then no change is needed.

The other major change is handling symlinks and hardlinks. Both are dereferenced in the archive creation because both types of links break TFE.

One positive result our of all of this work is that the giant tar case statement went away.

Other changes:

* Don't package .terraform/terraform.tfstate. It was done in legacy but no justification for including it could be found.
* Check for more commands that were being assumed
* Implement migration without VCS settings
* Clean up and consistently return instead of exit for more clarity in what's happening
* Style adjustments in checking for command success
* Fixes incorrect exit codes
* Fixes an error where sensitive values were not being redacted upon update
* Fixes an error where a path given for -var-file, such as some_dir/the_var_file was resulting in an error (awk couldn't find the file to process it).
* Resolves #8 and #12 .